+++ modules/sap_hana/main.tf
@@ -29,6 +29,14 @@ resource "google_compute_disk" "gcp_sap_hana_sd_0" {
   type    = var.disk_type_0
   zone    = var.zone
   size    = var.pd_ssd_size != "" ? var.pd_ssd_size : module.sap_hana.diskSize
+
+  # Add the disk_encryption_key block only if a pd_kms_key was provided
+  dynamic "disk_encryption_key" {
+    for_each = var.pd_kms_key != null ? [""] : []
+    content {
+      kms_key_self_link = var.pd_kms_key
+    }
+  }
 }
 
 resource "google_compute_disk" "gcp_sap_hana_sd_1" {
@@ -37,6 +45,14 @@ resource "google_compute_disk" "gcp_sap_hana_sd_1" {
   type    = var.disk_type_1
   zone    = var.zone
   size    = var.pd_hdd_size != "" ? var.pd_hdd_size : module.sap_hana.diskSize
+
+  # Add the disk_encryption_key block only if a pd_kms_key was provided
+  dynamic "disk_encryption_key" {
+    for_each = var.pd_kms_key != null ? [""] : []
+    content {
+      kms_key_self_link = var.pd_kms_key
+    }
+  }
 }
 
 resource "google_compute_address" "gcp_sap_hana_ip" {
@@ -59,7 +75,8 @@ resource "google_compute_instance" "gcp_sap_hana" {
   }
 
   boot_disk {
-    auto_delete = var.autodelete_disk
+    auto_delete       = var.autodelete_disk
+    kms_key_self_link = var.pd_kms_key
 
     initialize_params {
       image = "projects/${var.linux_image_project}/global/images/family/${var.linux_image_family}"